requirements.txt
# Dev environment
pip
autopep8

# FastAPI Backend
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2==2.9.9
pgvector==0.2.5
asyncpg==0.29.0

# Environment management
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# CORS handling
corsheaders==4.3.1

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0